+++ integrations/chroma/src/haystack_integrations/document_stores/chroma/utils.py
@@ -2,6 +2,8 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+from typing import Any
+
 from chromadb.api.types import EmbeddingFunction
 from chromadb.utils.embedding_functions import (
     CohereEmbeddingFunction,
@@ -34,7 +36,7 @@ FUNCTION_REGISTRY = {
 }
 
 
-def get_embedding_function(function_name: str, **kwargs) -> EmbeddingFunction:
+def get_embedding_function(function_name: str, **kwargs: Any) -> EmbeddingFunction:
     """Load an embedding function by name.
 
     :param function_name: the name of the embedding function.